feat: StoreInfo 객체 생성할 때 위치 정보 포함

+++ src/main/java/com/example/demo/store/dto/StoreInfoDTO.java
@@ -9,6 +9,6 @@ public record StoreInfoDTO(
     String id,
     String placeName,
     List<CardBenefitDTO> availableCards,
-    Integer lat,
-    Integer lng
+    Double lat,
+    Double lng
 ) {}
